requirements.txt
psycopg2-binary>=2.9.9
pymongo>=4.5.0
rich>=13.5.2
matplotlib>=3.8.0
numpy>=1.26.0

